introduce VacationType interface
