+++ src/test/java/org/synyx/urlaubsverwaltung/absence/web/AbsenceOverviewViewControllerTest.java
@@ -17,6 +17,7 @@ import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilde
 import org.synyx.urlaubsverwaltung.absence.AbsencePeriod;
 import org.synyx.urlaubsverwaltung.absence.AbsenceService;
 import org.synyx.urlaubsverwaltung.absence.DateRange;
+import org.synyx.urlaubsverwaltung.application.vacationtype.ProvidedVacationType;
 import org.synyx.urlaubsverwaltung.application.vacationtype.VacationCategory;
 import org.synyx.urlaubsverwaltung.application.vacationtype.VacationType;
 import org.synyx.urlaubsverwaltung.application.vacationtype.VacationTypeService;
@@ -892,7 +893,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -963,7 +964,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, absenceStatus, 1L, false);
@@ -1034,7 +1035,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -1116,7 +1117,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -1187,7 +1188,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, absenceStatus, 1L, false);
@@ -1258,7 +1259,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -1338,7 +1339,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, VacationCategory.OTHER, "other", false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, absenceStatus, 1L, false);
@@ -1407,7 +1408,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).category(OTHER).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).category(OTHER).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, VacationCategory.OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -1483,7 +1484,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(other, 1L, absenceStatus, 1L, false);
@@ -1552,7 +1553,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, absenceStatus, 1L, false);
@@ -1628,7 +1629,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, absenceStatus, 1L, false);
@@ -1698,7 +1699,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -1776,7 +1777,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, "other", false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, absenceStatus, 1L, false);
@@ -1845,7 +1846,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -1921,7 +1922,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(other, 1L, absenceStatus, 1L, false);
@@ -1990,7 +1991,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, absenceStatus, 1L, false);
@@ -2066,7 +2067,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, absenceStatus, 1L, false);
@@ -2136,7 +2137,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, absenceStatus, 1L, false);
@@ -2209,7 +2210,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2280,7 +2281,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2351,7 +2352,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2431,7 +2432,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2499,7 +2500,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2574,7 +2575,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(other, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2641,7 +2642,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2715,7 +2716,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2783,7 +2784,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -2855,7 +2856,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -2926,7 +2927,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -2997,7 +2998,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -3076,7 +3077,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -3144,7 +3145,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -3219,7 +3220,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(other, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -3286,7 +3287,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordNoonVacation noon = new AbsencePeriod.RecordNoonVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -3360,7 +3361,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
         when(departmentService.getDepartmentsPersonHasAccessTo(signedInUser)).thenReturn(List.of(department));
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(other, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -3428,7 +3429,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(VacationType.builder().id(1L).color(CYAN).build()));
+        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(ProvidedVacationType.builder().id(1L).color(CYAN).build()));
 //        when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(new VacationType(1L, true, OTHER, null, false, false, CYAN, false)));
 
         final AbsencePeriod.RecordMorningVacation morning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED_CANCELLATION_REQUESTED, 1L, false);
@@ -4118,7 +4119,7 @@ class AbsenceOverviewViewControllerTest {
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordMorningVacation vacationMorning = new AbsencePeriod.RecordMorningVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -4193,7 +4194,7 @@ class AbsenceOverviewViewControllerTest {
         department.setMembers(List.of(person));
         when(departmentService.getNumberOfDepartments()).thenReturn(1L);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(ORANGE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(ORANGE).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final AbsencePeriod.RecordNoonVacation vacationNoon = new AbsencePeriod.RecordNoonVacation(person, 1L, AbsencePeriod.AbsenceStatus.ALLOWED, 1L, false);
@@ -4342,7 +4343,7 @@ class AbsenceOverviewViewControllerTest {
         person.setEmail("springfield@example.org");
         when(personService.getSignedInUser()).thenReturn(person);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(CYAN).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(CYAN).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final var department = department();
@@ -4492,7 +4493,7 @@ class AbsenceOverviewViewControllerTest {
         person.setEmail("springfield@example.org");
         when(personService.getSignedInUser()).thenReturn(person);
 
-        final VacationType vacationType = VacationType.builder().id(1L).color(CYAN).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).color(CYAN).build();
         when(vacationTypeService.getAllVacationTypes()).thenReturn(List.of(vacationType));
 
         final var department = department();
@@ -4571,7 +4572,7 @@ class AbsenceOverviewViewControllerTest {
 
         when(vacationTypeService.getAllVacationTypes()).thenReturn(
             List.of(
-                VacationType.builder().id(1L).color(CYAN).build()
+                ProvidedVacationType.builder().id(1L).color(CYAN).build()
 //                new VacationType(1L, true, null, "", false, false, CYAN, true)
             )
         );
@@ -4655,7 +4656,7 @@ class AbsenceOverviewViewControllerTest {
 
         when(vacationTypeService.getAllVacationTypes()).thenReturn(
             List.of(
-                VacationType.builder().id(1L).color(CYAN).build()
+                ProvidedVacationType.builder().id(1L).color(CYAN).build()
 //                new VacationType(1L, true, null, "", false, false, CYAN, true)
             )
         );
@@ -4732,7 +4733,7 @@ class AbsenceOverviewViewControllerTest {
 
         when(vacationTypeService.getAllVacationTypes()).thenReturn(
             List.of(
-                VacationType.builder().id(1L).color(CYAN).build()
+                ProvidedVacationType.builder().id(1L).color(CYAN).build()
 //                new VacationType(1L, true, null, "", false, false, CYAN, true)
             )
         );
@@ -4810,8 +4811,8 @@ class AbsenceOverviewViewControllerTest {
 
         when(vacationTypeService.getAllVacationTypes())
             .thenReturn(List.of(
-                VacationType.builder().id(1L).color(ORANGE).build(),
-                VacationType.builder().id(2L).color(CYAN).build()
+                ProvidedVacationType.builder().id(1L).color(ORANGE).build(),
+                ProvidedVacationType.builder().id(2L).color(CYAN).build()
 //                new VacationType(1L, true, null, "", false, false, ORANGE, false),
 //                new VacationType(2L, true, null, "", false, false, CYAN, false)
             ));
@@ -4896,8 +4897,8 @@ class AbsenceOverviewViewControllerTest {
 
         when(vacationTypeService.getAllVacationTypes())
             .thenReturn(List.of(
-                VacationType.builder().id(1L).color(ORANGE).build(),
-                VacationType.builder().id(2L).color(CYAN).build()
+                ProvidedVacationType.builder().id(1L).color(ORANGE).build(),
+                ProvidedVacationType.builder().id(2L).color(CYAN).build()
 //                new VacationType(1L, true, null, "", false, false, ORANGE, false),
 //                new VacationType(2L, true, null, "", false, false, CYAN, false)
             ));
@@ -4973,8 +4974,8 @@ class AbsenceOverviewViewControllerTest {
 
         when(vacationTypeService.getAllVacationTypes())
             .thenReturn(List.of(
-                VacationType.builder().id(1L).color(ORANGE).build(),
-                VacationType.builder().id(2L).color(CYAN).build()
+                ProvidedVacationType.builder().id(1L).color(ORANGE).build(),
+                ProvidedVacationType.builder().id(2L).color(CYAN).build()
 //                new VacationType(1L, true, null, "", false, false, ORANGE, false),
 //                new VacationType(2L, true, null, "", false, false, CYAN, false)
             ));
