+++ src/test/java/org/synyx/urlaubsverwaltung/application/application/ApplicationForLeaveFormValidatorTest.java
@@ -11,6 +11,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.validation.Errors;
 import org.synyx.urlaubsverwaltung.application.settings.ApplicationSettings;
+import org.synyx.urlaubsverwaltung.application.vacationtype.ProvidedVacationType;
 import org.synyx.urlaubsverwaltung.application.vacationtype.VacationType;
 import org.synyx.urlaubsverwaltung.application.vacationtype.VacationTypeService;
 import org.synyx.urlaubsverwaltung.overlap.OverlapCase;
@@ -449,7 +450,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
         when(calculationService.checkApplication(any(Application.class))).thenReturn(TRUE);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(HOLIDAY).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(HOLIDAY).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -476,7 +477,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(UNPAIDLEAVE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(UNPAIDLEAVE).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -503,7 +504,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -530,7 +531,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(SPECIALLEAVE).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(SPECIALLEAVE).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -598,7 +599,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(errors.hasErrors()).thenReturn(FALSE);
         when(workingTimeService.getWorkingTime(any(Person.class), any(LocalDate.class))).thenReturn(Optional.of(createWorkingTime()));
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(HOLIDAY).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(HOLIDAY).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -628,7 +629,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(errors.hasErrors()).thenReturn(FALSE);
         when(workingTimeService.getWorkingTime(any(Person.class), any(LocalDate.class))).thenReturn(Optional.of(createWorkingTime()));
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(HOLIDAY).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(HOLIDAY).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -683,7 +684,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -711,11 +712,11 @@ class ApplicationForLeaveFormValidatorTest {
         when(calculationService.checkApplication(any(Application.class))).thenReturn(TRUE);
 
         when(vacationTypeService.getById(1L))
-            .thenReturn(Optional.of(VacationType.builder().id(1L).messageKey("message_key_1").category(HOLIDAY).build()));
+            .thenReturn(Optional.of(ProvidedVacationType.builder().id(1L).messageKey("message_key_1").category(HOLIDAY).build()));
         when(vacationTypeService.getById(2L))
-            .thenReturn(Optional.of(VacationType.builder().id(2L).messageKey("message_key_2").category(SPECIALLEAVE).build()));
+            .thenReturn(Optional.of(ProvidedVacationType.builder().id(2L).messageKey("message_key_2").category(SPECIALLEAVE).build()));
         when(vacationTypeService.getById(3L))
-            .thenReturn(Optional.of(VacationType.builder().id(3L).messageKey("message_key_3").category(UNPAIDLEAVE).build()));
+            .thenReturn(Optional.of(ProvidedVacationType.builder().id(3L).messageKey("message_key_3").category(UNPAIDLEAVE).build()));
 
         Consumer<ApplicationForLeaveFormVacationTypeDto> assertHoursNotMandatory = (vacationTypeDto) -> {
             appForm.setVacationType(vacationTypeDto);
@@ -758,7 +759,7 @@ class ApplicationForLeaveFormValidatorTest {
         settings.getOvertimeSettings().setOvertimeActive(false);
         when(settingsService.getSettings()).thenReturn(settings);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -784,7 +785,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -811,7 +812,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -866,7 +867,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workingTimeService.getWorkingTime(any(Person.class), any(LocalDate.class))).thenReturn(Optional.of(createWorkingTime()));
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -893,7 +894,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workingTimeService.getWorkingTime(any(Person.class), any(LocalDate.class))).thenReturn(Optional.of(createWorkingTime()));
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -920,7 +921,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -946,7 +947,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -973,7 +974,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(overlapService.checkOverlap(any(Application.class))).thenReturn(NO_OVERLAPPING);
         when(overtimeService.getLeftOvertimeForPerson(any(Person.class))).thenReturn(Duration.ofHours(10));
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -999,7 +1000,7 @@ class ApplicationForLeaveFormValidatorTest {
 
         setupOvertimeSettings();
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -1122,7 +1123,7 @@ class ApplicationForLeaveFormValidatorTest {
         when(workDaysCountService.getWorkDaysCount(any(DayLength.class), any(LocalDate.class), any(LocalDate.class), any(Person.class))).thenReturn(ONE);
         when(calculationService.checkApplication(any(Application.class))).thenReturn(TRUE);
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(HOLIDAY).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(HOLIDAY).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -1385,7 +1386,7 @@ class ApplicationForLeaveFormValidatorTest {
 
     private void overtimeMinimumTest(BigDecimal hours) {
 
-        final VacationType vacationType = VacationType.builder().id(1L).category(OVERTIME).build();
+        final VacationType vacationType = ProvidedVacationType.builder().id(1L).category(OVERTIME).build();
         when(vacationTypeService.getById(1L)).thenReturn(Optional.of(vacationType));
 
         final ApplicationForLeaveFormVacationTypeDto vacationTypeDto = new ApplicationForLeaveFormVacationTypeDto();
@@ -1413,6 +1414,6 @@ class ApplicationForLeaveFormValidatorTest {
     }
 
     private VacationType anyVacationType() {
-        return VacationType.builder().id(1L).category(HOLIDAY).messageKey("message_key").build();
+        return ProvidedVacationType.builder().id(1L).category(HOLIDAY).messageKey("message_key").build();
     }
 }
